# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mediasearch",
    "summary": "EmbeddingGemma 2 を使った録画映像・音声・画像のマルチモーダル検索 API(開発用)",
    "description": "\n録画映像・音声・画像を取り込み、**文章・画像・音声**で該当する場面を探すための API です。\n\n- 取り込み(`/api/ingest/*`)はバックグラウンドのジョブで行い、応答の `job_id` で進み具合を確認します\n  (`GET /api/jobs/{job_id}`)。\n- 検索(`/api/search/*`)は、スコア(コサイン類似度)の高い順に窓(時間の区切り、または静止画 1 枚)を返します。\n- 元ファイルは `GET /api/media/{source_id}`(Range 対応)、サムネイルは `GET /api/thumb/{window_id}` で取得します。\n- エラーは `{\"detail\": \"...\"}` の形で返します(説明は日本語)。\n\n**認証**: サーバに `API_TOKEN` が設定されている場合、`Authorization: Bearer <トークン>` か\n`X-API-Key: <トークン>` が必要です\n(右上の Authorize から設定できます)。設定されていない場合は不要です。\n\nAI エージェントからは、同じ機能を MCP(`/mcp`)で利用できます。\n開発者向けの説明は docs/api.md と docs/mcp.md を参照してください。\n",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/compute/info": {
      "get": {
        "tags": [
          "compute"
        ],
        "summary": "推論側の情報",
        "operationId": "info_compute_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeInfo"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "取り込み元・ジョブ"
        ],
        "summary": "ジョブの状態",
        "description": "取り込みの進み具合(progress / total)と状態(queued / running / done / failed)を返します。",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/media/{source_id}": {
      "get": {
        "tags": [
          "配信"
        ],
        "summary": "元ファイル(Range 対応)",
        "description": "元の動画・音声・画像を返します。Range リクエストに対応するため、動画は `#t=秒` で頭出し再生できます。",
        "operationId": "get_media_api_media__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "元のファイル(Content-Type はファイルの種類に合わせる。Range 指定時は 206)",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
    "/api/thumb/{window_id}": {
      "get": {
        "tags": [
          "配信"
        ],
        "summary": "窓のサムネイル(JPEG)",
        "operationId": "get_thumb_api_thumb__window_id__get",
        "parameters": [
          {
            "name": "window_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Window Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "長辺 320px の JPEG",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "認証が必要(API_TOKEN を設定している場合)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          },
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComputeInfo": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "dims": {
            "type": "integer",
            "title": "Dims"
          },
          "dtype": {
            "type": "string",
            "title": "Dtype"
          },
          "device": {
            "type": "string",
            "title": "Device"
          },
          "accelerator": {
            "type": "string",
            "title": "Accelerator"
          },
          "modalities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modalities"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "model_id",
          "dims",
          "dtype",
          "device",
          "accelerator",
          "modalities",
          "backend"
        ],
        "title": "ComputeInfo"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "description": "エラーの内容(日本語の説明。入力の検証エラーでは項目ごとの一覧)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "examples": [
          {
            "detail": "取り込み元が見つかりません"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "description": "処理済みの窓の数"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "窓の総数"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "source_id",
          "status",
          "progress",
          "total",
          "created_at"
        ],
        "title": "Job"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "API_TOKEN の値"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API_TOKEN の値"
      }
    }
  },
...
}